chore: move AccountContext folder
